requirements.txt
Django==2.1.4
Pillow==7.0.0
awesome-slugify==1.6.5
redis==3.4.1
markdown==3.2.1
django-braces==1.14.0
